+++ apps/admin/src/main.tsx
@@ -6,10 +6,17 @@ const bootstrap = async () => {
   const enableE2EMSW = (window as Window & { __ENABLE_E2E_MSW__?: boolean }).__ENABLE_E2E_MSW__ === true;
   if (enableE2EMSW) {
     const { worker } = await import('./mocks/browser');
-    await worker.start({
-      onUnhandledRequest: 'bypass',
-      serviceWorker: { url: '/mockServiceWorker.js' },
-    });
+    void worker
+      .start({
+        onUnhandledRequest: 'bypass',
+        serviceWorker: { url: '/mockServiceWorker.js' },
+      })
+      .then(() => {
+        window.__E2E_MSW_READY__ = true;
+      })
+      .catch(() => {
+        window.__E2E_MSW_READY__ = false;
+      });
   }
 
   ReactDOM.createRoot(document.getElementById('root')!).render(<App />);
